Created a Model calls to hold creates parameter. Changed the way the program unpacks parameters. Program notifies when it succeed or fail creating a file.

+++ FileManipulatorApp/FileManipulator/FileManager.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using FileManipulator.Models;
+using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
@@ -9,25 +10,42 @@ namespace FileManipulator
 {
     internal static class FileManager
     {
+        
         // string -create = "", string -dest = ".", int -amount = 1, string -append = "timestamp", bool -overwrite = false
         public static string Create(string[] args)
         {
-            Dictionary<string, string> parameters = SortParameters(args);
+            CreateModel parameter = SortParameters(args);
 
-            string destination = GetCurrentPath(parameters["-dest"]);
-            string fullFileDestination = Path.Join(destination, parameters["-name"]);
+            string destination = GetCurrentPath(parameter.Destination);
 
             if (Directory.Exists(destination))
             {
-                File.Create(fullFileDestination);
-                Console.WriteLine(fullFileDestination);
-                return "File Created"; ;
+                if (!String.IsNullOrWhiteSpace(parameter.Name))
+                {
+                    string fullFileDestination = Path.Join(destination, parameter.Name);
+                    if (File.Exists(fullFileDestination))
+                    {
+                        if (parameter.OverWrite)
+                        {
+                            File.Create(fullFileDestination);
+                            return "File Created";
+                        }
+                        else
+                        {
+                            return $"File {fullFileDestination} Already Exist";
+                        }
+                    } else
+                    {
+                        File.Create(fullFileDestination);
+                        return $"File Created at: {fullFileDestination}";
+                    }
+                }
             }
             else
             {
                 return "";
             }
-
+            return "";
         }
 
         private static string GetCurrentPath(string destination)
@@ -40,39 +58,39 @@ namespace FileManipulator
             return destination;
         }
 
-        private static Dictionary<string, string> SortParameters(string[] parameters)
+        private static CreateModel SortParameters(string[] parameters)
         {
-            Dictionary<string, string> returnValues = new();
-
+            CreateModel createModel = new CreateModel();
             foreach (var parameter in parameters)
             {
-                switch (parameter.ToLower())
-                {
-                    case "-dest":
-                        int indexValue = Array.IndexOf(parameters, "-dest");
-                        returnValues.Add(parameters[indexValue], parameters[indexValue + 1]);
-                        break;
-                    case "-name":
-                        indexValue = Array.IndexOf(parameters, "-name");
-                        returnValues.Add(parameters[indexValue], parameters[indexValue + 1]);
-                        break;
-                    case "-amount":
-                        indexValue = Array.IndexOf(parameters, "-amount");
-                        returnValues.Add(parameters[indexValue], parameters[indexValue + 1]);
-                        break;
-                    case "-append":
-                        indexValue = Array.IndexOf(parameters, "-append");
-                        returnValues.Add(parameters[indexValue], parameters[indexValue + 1]);
-                        break;
-                    case "-overwrite":
-                        indexValue = Array.IndexOf(parameters, "-overwrite");
-                        returnValues.Add(parameters[indexValue], parameters[indexValue + 1]);
-                        break;
-                    default:
-                        break;
+                if (parameter.Contains("-") && parameter.ToLower() != "-create") { 
+                    int index = Array.IndexOf(parameters, parameter);
+                    switch (parameter.ToLower())
+                    {
+                        case "-name":
+                            createModel.Name = parameters[index + 1];
+                            break;
+                        case "-dest":
+                            createModel.Destination = parameters[index + 1];
+                            break;
+                        case "-amount":
+                            createModel.Amount = Convert.ToInt32(parameters[index + 1]);
+                            break;
+                        case "-append":
+                            createModel.Append = parameters[index + 1];
+                            break;
+                        case "-overwrite":
+                            createModel.OverWrite = parameters[index + 1] == "true" ?  true : false;
+                            break;
+                        default:
+                            break;
+                    }
                 }
             }
-            return returnValues;
+
+
+
+            return createModel;
         }
     }
 }
